requirements.txt
python-dotenv~=1.1.1
requests
openai~=1.92.3
